streamline everything, fixes to come soon

+++ hdfs_test/__main__.py
@@ -20,26 +20,44 @@ from timeit import default_timer as timer
 
 # Parsing script arguments
 parser = argparse.ArgumentParser(description="Test writing or reading throughput of HDFS")
-parser.add_argument('-u', '--URL', help='URL of the dataset to use for testing, the file must be a single csv '
-                                        'in a zip archive, if this parameter is not specified the dataset is assumed '
-                                        'to have been downloaded already (default: blank)')
-parser.add_argument('-z', '--zip', help="Says what unpacker to use, zip or tar (default: zip)")
+parser.add_argument('-fr', '--firstRun', help="With this parameter the script will prepare Hdfs environment for testing",
+                    action="store_true")
+parser.add_argument('-u', '--URL', help='URL of the dataset to use for testing, the file must be one or more CSVs '
+                                        'in a zip or tar.gz archive or a plain csv, if this parameter is not specified the dataset, it is assumed '
+                                        'to have been downloaded already and present in dataset folder(default: blank)')
+parser.add_argument('-z', '--zip', help="Says what unpacker to use, zip, tar or none (default: zip)")
 parser.add_argument('-l', '--load', help="Use this parameter to load the dataset to hdfs",
                     action="store_true")
-parser.add_argument('-r', '--reading', help="With this parameter the script will test the reading throughput of the HDFS"
-                                            " instead of the default writing", action="store_true")
+parser.add_argument('-t', '--testing', help="With this parameter the script will test the reading and writing throughput of the HDFS", action="store_true")
 parser.add_argument('-f', '--flush', help="With this parameter the script will only clean up the HDFS",
                     action="store_true")
+
 parser.add_argument('-fa', '--flushAll', help="With this parameter the script will clean up the HDFS and local files",
                     action="store_true")
-parser.set_defaults(URL='', zip='zip')
+parser.add_argument('-ti', '--testIterations',	help='Number of iterations done for testing (default: 1')				
+parser.add_argument('-sa', '--sparkArguments', help='The parameters to be sent to spark (default: "--master yarn --num-executors 1 --executor-memory 1G")')
+parser.set_defaults(URL='', zip='zip', testIterations = '1', sparkArguments='--master yarn --num-executors 1 --executor-memory 1G')
 args = parser.parse_args()
 
+firstRun = args.firstRun
 URL = args.URL
-reading = args.reading
+testing = args.testing
 flush = args.flush
 flushAll = args.flushAll
 loading = args.load
+ti = args.testIterations
+sparkArguments = args.sparkArguments
+
+if firstRun:
+	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -mkdir /user/", shell=True)
+	process.wait()
+	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -mkdir /user/admin/", shell=True)
+	process.wait()
+	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -mkdir /user/admin/testing/", shell=True)
+	process.wait()
+	process = subprocess.Popen("chmod 755 /root/", shell=True)
+	process.wait()
+
 
 if flushAll:
 	process = subprocess.Popen("rm -f test.csv", shell=True)
@@ -57,60 +75,44 @@ if URL != '':
 		process.wait()
 		process = subprocess.Popen("wget %s -O test.zip" % URL, shell=True)
 		process.wait()
-		process = subprocess.Popen("unzip test.zip", shell=True)
+		process = subprocess.Popen("unzip test.zip -d ./dataset/", shell=True)
 		process.wait()
-		process = subprocess.Popen("mv *.csv test.csv", shell=True)
-		process.wait()
+		#process = subprocess.Popen("mv *.csv test.csv", shell=True)
+		#process.wait()
 		process = subprocess.Popen("rm -f test.zip", shell=True)
 		process.wait()
-	else:
+	elif zip == 'tar':
 		process = subprocess.Popen("yum install tar", shell=True)
 		process.wait()
 		process = subprocess.Popen("wget %s -O test.tar.gz" % URL, shell=True)
 		process.wait()
-		process = subprocess.Popen("tar -xvzf test.tar.gz", shell=True)
+		process = subprocess.Popen("tar -xzf test.tar.gz -C ./dataset/", shell=True)
 		process.wait()
-		process = subprocess.Popen("mv *.csv test.csv", shell=True)
-		process.wait()
+		#process = subprocess.Popen("mv *.csv test.csv", shell=True)
+		#process.wait()
 		process = subprocess.Popen("rm -f test.tar.gz", shell=True)
 		process.wait()
+	else:
+		process = subprocess.Popen("wget %s -O ./dataset/test.csv" % URL, shell=True)
 
 if loading:
-	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -put -f test.csv /user/admin/testing/test.csv", shell=True)
+	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -copyFromLocal -f ./dataset /user/admin/testing/dataset", shell=True)
 	process.wait()
-		
+	
 times = []
-'''
-if reading:
-	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -put -f test.csv /user/admin/testing/test.csv", shell=True)
-	process.wait()
-	process = subprocess.Popen("rm -f test.csv", shell=True)
-	process.wait()
+
+if testing:
+
 	
-	for x in range(0, 9):
+	for x in range(0, int(ti)):
 		start = timer()
-		process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -get /user/admin/testing/test.csv test.csv", shell=True)
+		process = subprocess.Popen("sudo -u hdfs spark-submit %s ./hdfs_test/rHdfsWRam.py" % sparkArguments, shell=True)
 		process.wait()
 		end = timer()
-		process = subprocess.Popen("rm -f test.csv", shell=True)
+		process = subprocess.Popen("sudo -u hdfs spark-submit %s ./hdfs_test/rRamWHdfs.py" % sparkArguments, shell=True)
 		process.wait()
 		times.append(end-start)
-	
-	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -get /user/admin/testing/test.csv test.csv", shell=True)
-	process.wait()
-	process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -rm -f -skipTrash /user/admin/testing/test.csv", shell=True)
-	process.wait()
-	
-else:
-	for x in range(0, 9):
-		start = timer()
-		process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -put -f test.csv /user/admin/testing/test.csv", shell=True)
-		process.wait()
-		end = timer()
-		process = subprocess.Popen("HADOOP_USER_NAME=hdfs hadoop fs -rm -f -skipTrash /user/admin/testing/test.csv", shell=True)
-		process.wait()
-		times.append(end-start)
+
 	
 print "Average Speed for %s GB: %s" \
         % (str(os.path.getsize("./test.csv")/pow(1024, 3)), str(sum(times)/len(times)))
-'''
